Move compatibility layer under 'compat' package.
