+++ src/test/java/com/github/dakusui/cmd/compat/CommandUtils.java
@@ -1,8 +1,10 @@
-package com.github.dakusui.cmd;
+package com.github.dakusui.cmd.compat;
 
+import com.github.dakusui.cmd.Cmd;
+import com.github.dakusui.cmd.Shell;
 import com.github.dakusui.cmd.exceptions.CommandException;
 import com.github.dakusui.cmd.exceptions.UnexpectedExitValueException;
-import com.github.dakusui.cmd.exceptions.CommandTimeoutException;
+import com.github.dakusui.cmd.compat.exceptions.CommandTimeoutException;
 import com.github.dakusui.cmd.exceptions.Exceptions;
 import com.github.dakusui.cmd.io.RingBufferedLineWriter;
