+++ src/test/java/com/github/dakusui/cmd/compat/exceptions/CommandTimeoutException.java
@@ -1,4 +1,6 @@
-package com.github.dakusui.cmd.exceptions;
+package com.github.dakusui.cmd.compat.exceptions;
+
+import com.github.dakusui.cmd.exceptions.CommandException;
 
 public class CommandTimeoutException extends CommandException {
   public CommandTimeoutException(Throwable t) {
